# apps/backend/src/main/resources/api.yaml
...
openapi: 3.0.0
info:
  title: StoreMe API
  version: "1"
servers:
  - url: http://localhost:8091
    description: Main API
...
paths:
...
  "/inventory/{inventoryId}/item-models":
    parameters:
      - in: path
        name: inventoryId
        schema:
          type: integer
        required: true
        description: Numeric of the current inventory
    get:
      tags:
        - ItemModel
      description: Get item models, supports the swagger-undocumented filter
      operationId: getItemModels
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemModelDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    post:
      tags:
        - ItemModel
      description: Create item model
      operationId: createItemModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemModelDTO"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemModelDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
  "/item-models/{id}":
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the resource to get
    get:
      tags:
        - ItemModel
      description: Get item model
      operationId: getItemModel
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemModelDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    put:
      tags:
        - ItemModel
      description: Update item model
      operationId: updateItemModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemModelDTO"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemModelDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    delete:
      tags:
        - ItemModel
      description: Delete item model
      operationId: deleteItemModel
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemModelDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
components:
  schemas:
    JSONErrorMessage:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
...
    ItemModelDTO:
      x-implements:
        - ch.heig.pdg.backend.dto.IDataTransferObject<ch.heig.pdg.backend.entities.ItemModel>
      required:
        - description
        - name
      properties:
        id:
          type: integer
        updated:
          type: string
        created:
          type: string
        name:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: integer
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "401":
      description: No authorization was provided or token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "409":
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "500":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"